VF-1662 Fix ios12 compatibility
Quoting upstream bugreport:
When creating a pipeline with audio on iOS, if the channel layout is not
kAudioChannelLayoutTag_UseChannelDescriptions, the pipeline will not be created.

For versions of iOS up to iOS 11, this is not an issue
as kAudioChannelLayoutTag_UseChannelDescriptions is the default channel layout.
However, starting in iOS 12, the channel layout is mandated by the hardware
of the device you are running. This value is also no longer editable in iOS 12.

Add support for stereo and dont fail when channel is not known

Apply patches from https://bugzilla.gnome.org/show_bug.cgi?id=796919